Add files via upload
Defines a simple program that helps users track stock prices and make investment decisions. It includes a menu with several options for users to choose from regarding their stocks. Creates a basic investment plan that suggests the best day to buy and sell based on the historical prices to maximize profit. The program loops through the menu options until the user chooses to quit.