Add files via upload
defines a Date class to represent and manipulate dates. The class includes methods to check if a year is a leap year, get the number of days, and copy the date. It can compare dates to check if one is before or after another and find the number of days between two dates. Dates are represented as strings in "MM/DD/YYYY" format with the __repr__ method, and equality between two dates is checked using the __eq__ method.