Add files via upload
This program creates random text based on transitions in the inputted text file. Reading the files it creates a dictionary of keys (words), where the values of the key is a list of words that follow the key in the original text file.  Resetting every time it reaches punctuation, the generate_text() fn randomly prints words based on the created dictionary.